+++ test/observables/states/states_combine_test.dart
@@ -3,7 +3,7 @@ import 'package:test/test.dart';
 import 'package:scopes/scopes.dart';
 
 import '../../test_gen/combine_test/combine_test.dart';
-import '../../toolbox/states_tester.dart';
+import '../shared/states_tester.dart';
 
 part 'states_combine_test.g.dart';
