+++ test/scopes/scopes/scope_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:test/test.dart';
 import 'package:scopes/scopes.dart';
 
-import '../../toolbox/mock_configurable.dart';
+import '../shared/mock_configurable.dart';
 
 void main() {
